# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: VoiceAsset Server API
  version: 0.22.0
  summary: Public REST contract for the VoiceAsset modular monolith.
  description: >-
    VoiceAsset provides local web sessions, immutable audio ingestion, durable
    transcription jobs, encrypted workspace-scoped ASR provider profiles,
    transcript retrieval, authenticated byte-range audio delivery, and bounded
    asynchronous waveform derivatives. Resource access is always scoped to the
    authenticated workspace.
  license:
    name: GNU Affero General Public License v3.0 or later
    identifier: AGPL-3.0-or-later
servers:
  - url: /
    description: Server hosting this contract
...
security: []
paths:
...
  /api/v1/assets:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      tags: [Assets]
      summary: List or search voice assets
      description: >-
        Requires `assets:read`. Results are workspace-scoped and ordered by
        immutable creation time, then identifier, both descending. The opaque
        cursor must be reused with the same complete filter set. Trashed assets
        are excluded unless `status=trashed` is requested explicitly.
      operationId: listAssets
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssetSearchQuery'
        - $ref: '#/components/parameters/AssetCollectionFilter'
        - $ref: '#/components/parameters/AssetTagFilter'
        - $ref: '#/components/parameters/AssetStatusFilter'
        - $ref: '#/components/parameters/AssetProviderFilter'
        - $ref: '#/components/parameters/AssetSpeakerFilter'
        - $ref: '#/components/parameters/AssetCreatedFrom'
        - $ref: '#/components/parameters/AssetCreatedBefore'
        - $ref: '#/components/parameters/ListLimit'
        - $ref: '#/components/parameters/ListCursor'
      responses:
        '200':
          $ref: '#/components/responses/AssetListResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
    post:
      tags: [Assets]
      summary: Create a draft voice asset
      description: >-
        Requires `assets:write`. Reusing an Idempotency-Key with the same
        normalized input returns the original asset and
        `Idempotency-Replayed: true`; different input returns 409. A
        cookie-authenticated request must send the configured Origin.
      operationId: createAsset
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '200':
          $ref: '#/components/responses/AssetReplayResponse'
        '201':
          $ref: '#/components/responses/AssetCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '413':
          $ref: '#/components/responses/RequestTooLargeResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
...
  /api/v1/assets/{asset_id}:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
      - $ref: '#/components/parameters/AssetID'
    get:
      tags: [Assets]
      summary: Get a voice asset
      description: Requires `assets:read`; an asset in another workspace is returned as 404.
      operationId: getAsset
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/AssetResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
    delete:
      tags: [Assets]
      summary: Move a voice asset to trash
      description: >-
        Requires `assets:write` and the exact resource-version ETag. The asset,
        immutable audio, transcripts, and metadata remain stored; normal reads
        hide the asset until it is restored. Successful changes are
        transactionally audited. Cookie authentication requires Origin.
      operationId: trashAsset
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/AssetIfMatch'
      responses:
        '200':
          $ref: '#/components/responses/AssetResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '428':
          $ref: '#/components/responses/PreconditionRequiredResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: voiceasset_session
      description: >-
        Opaque HttpOnly session cookie. Unsafe cookie-authenticated operations
        also require Origin to equal the configured public origin.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: voiceasset_refresh
      description: >-
        Opaque HttpOnly refresh cookie scoped to `/api/v1/auth`. It is accepted
        only by the rotation endpoint together with the exact configured Origin.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: VoiceAsset opaque token
      description: >-
        Opaque session token beginning with `va_`, or scoped API key beginning
        with `va_pat_`. If both cookie and bearer credentials are sent, they
        must contain the same token. API keys cannot revoke web sessions.
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; oversized values are replaced.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    Origin:
      name: Origin
      in: header
      required: false
      description: >-
        Browser origin. Required and exact for unsafe cookie-authenticated
        operations; optional for bearer authentication.
      schema:
        type: string
        format: uri
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated retry key scoped to the operation and workspace.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    AssetID:
      name: asset_id
      in: path
      required: true
      description: Voice asset identifier.
      schema:
        type: string
        format: uuid
...
    AssetSearchQuery:
      name: q
      in: query
      required: false
      description: >-
        PostgreSQL full-text terms matched against asset titles and segments in
        each Transcript's latest immutable Revision. Literal matching is retained
        for titles and languages that are not tokenized by the simple dictionary.
      schema:
        type: string
        maxLength: 200
    AssetCollectionFilter:
      name: collection_id
      in: query
      required: false
      description: Return only assets assigned to this workspace collection.
      schema:
        type: string
        format: uuid
    AssetTagFilter:
      name: tag_id
      in: query
      required: false
      description: Return only assets assigned to this workspace tag.
      schema:
        type: string
        format: uuid
    AssetStatusFilter:
      name: status
      in: query
      required: false
      description: Return only assets in this lifecycle or processing status.
      schema:
        type: string
        enum: [draft, uploading, processing, ready, failed, trashed]
    AssetProviderFilter:
      name: provider_id
      in: query
      required: false
      description: Return only assets with a Raw ASR Revision produced by this provider.
      schema:
        type: string
        enum: [mock_asr, aliyun_asr, tencent_asr]
    AssetSpeakerFilter:
      name: speaker
      in: query
      required: false
      description: >-
        Case-insensitive exact Speaker label in a latest immutable Transcript
        Revision. When q is present, Segment text hits are narrowed to this Speaker.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    AssetCreatedFrom:
      name: created_from
      in: query
      required: false
      description: Inclusive lower bound for immutable asset creation time.
      schema:
        type: string
        format: date-time
    AssetCreatedBefore:
      name: created_before
      in: query
      required: false
      description: Exclusive upper bound for immutable asset creation time.
      schema:
        type: string
        format: date-time
    ListLimit:
      name: limit
      in: query
      required: false
      description: Maximum number of assets to return; defaults to 20.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ListCursor:
      name: cursor
      in: query
      required: false
      description: Opaque continuation cursor returned by the preceding page.
      schema:
        type: string
        minLength: 1
        maxLength: 1024
...
    AssetIfMatch:
      name: If-Match
      in: header
      required: true
      description: Exact quoted asset version returned in ETag.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
      example: '"3"'
...
  headers:
    XAPIVersion:
      description: Stable API namespace served by this process.
      schema:
        type: string
        const: v1
    XRequestID:
      description: Caller-provided or server-generated request correlation ID.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    XServerVersion:
      description: Semantic server build version.
      schema:
        type: string
        minLength: 1
    Location:
      description: Relative URL of the created or completed resource.
      schema:
        type: string
    ETag:
      description: Quoted resource version or lowercase SHA-256 entity tag.
      schema:
        type: string
    IdempotencyReplayed:
      description: Present with value `true` when an earlier result is replayed.
      schema:
        type: string
        const: 'true'
...
    WWWAuthenticate:
      description: Bearer authentication challenge.
      schema:
        type: string
        const: 'Bearer realm="voiceasset"'
...
  responses:
...
    AssetCreatedResponse:
      description: New draft asset.
      headers:
        Location:
          $ref: '#/components/headers/Location'
        ETag:
          $ref: '#/components/headers/ETag'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Asset'
    AssetReplayResponse:
      description: Previously created asset replayed.
      headers:
        Location:
          $ref: '#/components/headers/Location'
        ETag:
          $ref: '#/components/headers/ETag'
        Idempotency-Replayed:
          $ref: '#/components/headers/IdempotencyReplayed'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Asset'
    AssetResponse:
      description: Workspace-scoped voice asset.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Asset'
...
    AssetListResponse:
      description: Stable page of workspace-scoped voice assets.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AssetList'
...
    BadRequestResponse:
      description: Invalid JSON, identifier, idempotency key, or operation input.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedResponse:
      description: Session credential is missing, malformed, expired, revoked, or ambiguous.
      headers:
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ForbiddenResponse:
      description: Origin validation failed or the authenticated principal lacks the required scope.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundResponse:
      description: Resource does not exist in the authenticated workspace.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowedResponse:
      description: The route exists but does not support this HTTP method.
      headers:
        Allow:
          description: Comma-separated methods accepted by the route.
          schema:
            type: string
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictResponse:
      description: Current resource, membership, idempotency input, upload state, or asset state conflicts with this request.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RequestTooLargeResponse:
      description: JSON or binary request body exceeds the operation limit.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalErrorResponse:
      description: Request could not be completed due to an internal failure.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailableResponse:
      description: A required application service is not configured or not ready.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PreconditionRequiredResponse:
      description: A required If-Match version was omitted.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateAssetRequest:
      type: object
      additionalProperties: false
      required: [title, language]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Trimmed title without control characters.
        language:
          $ref: '#/components/schemas/Language'
...
    AssetList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Asset'
        next_cursor:
          type: string
          minLength: 1
          maxLength: 1024
...
    Language:
      type: string
      description: BCP 47-like language tag or `und` for undetermined.
      pattern: '^(?:und|[A-Za-z]{2,3}(?:-[A-Za-z0-9]{2,8})*)$'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
        request_id:
          type: string
          minLength: 1
          maxLength: 200
...